Have all symbols available for the expressionEvaluator Separate utility to parse wiki symbols to their html representation Removed unused imports
